+++ source/android/adaptivecards/src/main/java/io/adaptivecards/renderer/Util.java
@@ -394,4 +394,13 @@ public final class Util {
         CopyInputProperties(baseInputElement, inputElement);
     }
 
+    public static long getViewId(View view)
+    {
+        if(view.getId() == View.NO_ID)
+        {
+            view.setId(View.generateViewId());
+        }
+        return view.getId();
+    }
+
 }
